+++ src/http/input/metadata/ContentTypeParser.ts
@@ -4,7 +4,6 @@ import { MetadataParser } from './MetadataParser';
 
 /**
  * Parser for the `content-type` header.
- * Currently only stores the media type and ignores other parameters such as charset.
  */
 export class ContentTypeParser extends MetadataParser {
   public async handle(input: { request: HttpRequest; metadata: RepresentationMetadata }): Promise<void> {