+++ test/unit/http/ldp/metadata/WebAclMetadataCollector.test.ts
@@ -5,9 +5,11 @@ import { WebAclMetadataCollector } from '../../../../../src/http/ldp/metadata/We
 import type { Operation } from '../../../../../src/http/Operation';
 import { BasicRepresentation } from '../../../../../src/http/representation/BasicRepresentation';
 import { RepresentationMetadata } from '../../../../../src/http/representation/RepresentationMetadata';
+import { IdentifierMap } from '../../../../../src/util/map/IdentifierMap';
 import { ACL, AUTH } from '../../../../../src/util/Vocabularies';
 
 describe('A WebAclMetadataCollector', (): void => {
+  const target = { path: 'http://example.com/foo' };
   let operation: Operation;
   let metadata: RepresentationMetadata;
   const writer = new WebAclMetadataCollector();
@@ -15,7 +17,7 @@ describe('A WebAclMetadataCollector', (): void => {
   beforeEach(async(): Promise<void> => {
     operation = {
       method: 'GET',
-      target: { path: 'http://test.com/foo' },
+      target,
       preferences: {},
       body: new BasicRepresentation(),
     };
@@ -23,27 +25,38 @@ describe('A WebAclMetadataCollector', (): void => {
     metadata = new RepresentationMetadata();
   });
 
+  it('adds no metadata if there is no target entry.', async(): Promise<void> => {
+    await expect(writer.handle({ metadata, operation })).resolves.toBeUndefined();
+    expect(metadata.quads()).toHaveLength(0);
+
+    operation.availablePermissions = new IdentifierMap();
+    await expect(writer.handle({ metadata, operation })).resolves.toBeUndefined();
+    expect(metadata.quads()).toHaveLength(0);
+  });
+
   it('adds no metadata if there are no permissions.', async(): Promise<void> => {
     await expect(writer.handle({ metadata, operation })).resolves.toBeUndefined();
     expect(metadata.quads()).toHaveLength(0);
 
-    operation.permissionSet = {};
+    operation.availablePermissions = new IdentifierMap([[ target, {}]]);
     await expect(writer.handle({ metadata, operation })).resolves.toBeUndefined();
     expect(metadata.quads()).toHaveLength(0);
   });
 
   it('adds no metadata if the method is wrong.', async(): Promise<void> => {
-    operation.permissionSet = { [CredentialGroup.public]: { read: true, write: false }};
+    operation.availablePermissions = new IdentifierMap(
+      [[ target, { [CredentialGroup.public]: { read: true, write: false }}]],
+    );
     operation.method = 'DELETE';
     await expect(writer.handle({ metadata, operation })).resolves.toBeUndefined();
     expect(metadata.quads()).toHaveLength(0);
   });
 
   it('adds corresponding metadata for all permissions present.', async(): Promise<void> => {
-    operation.permissionSet = {
+    operation.availablePermissions = new IdentifierMap([[ target, {
       [CredentialGroup.agent]: { read: true, write: true, control: false } as AclPermission,
       [CredentialGroup.public]: { read: true, write: false },
-    };
+    }]]);
     await expect(writer.handle({ metadata, operation })).resolves.toBeUndefined();
     expect(metadata.quads()).toHaveLength(3);
     expect(metadata.getAll(AUTH.terms.userMode)).toEqualRdfTermArray([ ACL.terms.Read, ACL.terms.Write ]);
@@ -51,10 +64,10 @@ describe('A WebAclMetadataCollector', (): void => {
   });
 
   it('ignores unknown modes.', async(): Promise<void> => {
-    operation.permissionSet = {
+    operation.availablePermissions = new IdentifierMap([[ target, {
       [CredentialGroup.agent]: { read: true, create: true },
       [CredentialGroup.public]: { read: true },
-    };
+    }]]);
     await expect(writer.handle({ metadata, operation })).resolves.toBeUndefined();
     expect(metadata.quads()).toHaveLength(2);
     expect(metadata.getAll(AUTH.terms.userMode)).toEqualRdfTermArray([ ACL.terms.Read ]);